+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/MapperSupportDaoImpl.java
@@ -3,7 +3,7 @@ package com.dwarfeng.fdr.impl.dao;
 import com.dwarfeng.fdr.impl.bean.entity.HibernateMapperSupport;
 import com.dwarfeng.fdr.stack.bean.entity.MapperSupport;
 import com.dwarfeng.fdr.stack.dao.MapperSupportDao;
-import com.dwarfeng.subgrade.impl.dao.HibernateBaseDao;
+import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
 import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
 import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
 import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
@@ -12,7 +12,6 @@ import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,46 +20,99 @@ import java.util.List;
 @Repository
 public class MapperSupportDaoImpl implements MapperSupportDao {
 
-    @Autowired
-    private HibernateBaseDao<StringIdKey, HibernateStringIdKey, MapperSupport, HibernateMapperSupport> baseDao;
-    @Autowired
-    private HibernateEntireLookupDao<MapperSupport, HibernateMapperSupport> entireLookupDao;
-    @Autowired
-    private HibernatePresetLookupDao<MapperSupport, HibernateMapperSupport> presetLookupDao;
+    private final HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, MapperSupport, HibernateMapperSupport>
+            batchBaseDao;
+    private final HibernateEntireLookupDao<MapperSupport, HibernateMapperSupport> entireLookupDao;
+    private final HibernatePresetLookupDao<MapperSupport, HibernateMapperSupport> presetLookupDao;
+
+    public MapperSupportDaoImpl(
+            HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, MapperSupport, HibernateMapperSupport>
+                    batchBaseDao,
+            HibernateEntireLookupDao<MapperSupport, HibernateMapperSupport> entireLookupDao,
+            HibernatePresetLookupDao<MapperSupport, HibernateMapperSupport> presetLookupDao
+    ) {
+        this.batchBaseDao = batchBaseDao;
+        this.entireLookupDao = entireLookupDao;
+        this.presetLookupDao = presetLookupDao;
+    }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public StringIdKey insert(MapperSupport element) throws DaoException {
-        return baseDao.insert(element);
+        return batchBaseDao.insert(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void update(MapperSupport element) throws DaoException {
-        baseDao.update(element);
+        batchBaseDao.update(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void delete(StringIdKey key) throws DaoException {
-        baseDao.delete(key);
+        batchBaseDao.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public boolean exists(StringIdKey key) throws DaoException {
-        return baseDao.exists(key);
+        return batchBaseDao.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public MapperSupport get(StringIdKey key) throws DaoException {
-        return baseDao.get(key);
+        return batchBaseDao.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsert(@SkipRecord List<MapperSupport> elements) throws DaoException {
+        return batchBaseDao.batchInsert(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdate(@SkipRecord List<MapperSupport> elements) throws DaoException {
+        batchBaseDao.batchUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        batchBaseDao.batchDelete(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        return batchBaseDao.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        return batchBaseDao.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<MapperSupport> batchGet(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        return batchBaseDao.batchGet(keys);
     }
 
     @Override