+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/dao/MapperSupportDao.java
@@ -2,7 +2,7 @@ package com.dwarfeng.fdr.stack.dao;
 
 import com.dwarfeng.fdr.stack.bean.entity.MapperSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.dao.BaseDao;
+import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
 import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
 import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
 
@@ -12,6 +12,6 @@ import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
  * @author DwArFeng
  * @since 1.6.0
  */
-public interface MapperSupportDao extends BaseDao<StringIdKey, MapperSupport>, EntireLookupDao<MapperSupport>,
+public interface MapperSupportDao extends BatchBaseDao<StringIdKey, MapperSupport>, EntireLookupDao<MapperSupport>,
         PresetLookupDao<MapperSupport> {
 }